WIP - refactoring
